Define multi-environment registry
This is a first step into defining a better structured for the global
registry we will need for the app to look up the supported chains and
tokens. The registry will either be fully dynamic in that requests will
be made for each need or we will generate a static file every X hours
based on the current state of the different moving pieces (supported
chains and bridges).