Disable fetching allowance every 3s
For some reason, *sometimes* it makes the TokenSpendApproval component
disappear. For instance, I set the amount to X, the component shows up,
then I change the amount to X+1 and it still shows but ~3s later it's
gone. Couldn't see the bug happen using console.log but disabling the
interval fixes it so I will do that for now.