fix: Only apply formatAmount on  max amount
I discovered that by always applying `formatAmount` on the input amount,
we are blocking the user from inputting certain values like `0` (it was
seen as `null` so it resulted in '') or `0,0` since it was formatted to
`0`, etc. So we should let the user input what they want freely and we
only apply `formatAmount` when handling the Max button action. I tested
this all around and all seems fine now.